+++ crates/server/src/types.rs
@@ -8,8 +8,8 @@ pub struct PitchLakeJobRequestParams {
     pub twap: (i64, i64),
     pub cap_level: (i64, i64),
     pub reserve_price: (i64, i64),
-    pub alpha: i64,
-    pub k: i64,
+    pub alpha: u128,
+    pub k: i128,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
